Create library package for common security classes
