+++ libraries/security/src/main/java/org/dataportabilityproject/security/DecrypterFactory.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dataportabilityproject.gateway.crypto;
+package org.dataportabilityproject.security;
 
 import com.google.common.base.Preconditions;
 import java.security.PrivateKey;
