+++ libraries/security/src/main/java/org/dataportabilityproject/security/Encrypter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dataportabilityproject.gateway.crypto;
+package org.dataportabilityproject.security;
 
 /**
  * Provides functionality to encrypt raw data.
