+++ libraries/security/src/main/java/org/dataportabilityproject/security/RsaSymmetricKeyGenerator.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dataportabilityproject.gateway.crypto;
+package org.dataportabilityproject.security;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
