+++ portability-gateway/build.gradle
@@ -36,6 +36,7 @@ jar {
 }
 
 dependencies {
+    compile project(':libraries:security')
     compile project(':portability-spi-cloud')
     compile project(':portability-spi-gateway')
     compile project(':portability-types-client')
