+++ portability-gateway/src/main/java/org/dataportabilityproject/gateway/reference/DataTransferHandler.java
@@ -33,8 +33,8 @@ import org.dataportabilityproject.gateway.ApiSettings;
 import org.dataportabilityproject.gateway.action.createjob.CreateJobAction;
 import org.dataportabilityproject.gateway.action.createjob.CreateJobActionRequest;
 import org.dataportabilityproject.gateway.action.createjob.CreateJobActionResponse;
-import org.dataportabilityproject.gateway.crypto.EncrypterFactory;
-import org.dataportabilityproject.gateway.crypto.SymmetricKeyGenerator;
+import org.dataportabilityproject.security.EncrypterFactory;
+import org.dataportabilityproject.security.SymmetricKeyGenerator;
 import org.dataportabilityproject.gateway.reference.ReferenceApiUtils.HttpMethods;
 import org.dataportabilityproject.spi.cloud.storage.JobStore;
 import org.dataportabilityproject.spi.cloud.types.JobAuthorization;
