+++ portability-gateway/src/main/java/org/dataportabilityproject/gateway/reference/Oauth2CallbackHandler.java
@@ -31,9 +31,9 @@ import java.util.Map;
 import java.util.UUID;
 import javax.crypto.SecretKey;
 import org.dataportabilityproject.gateway.ApiSettings;
-import org.dataportabilityproject.gateway.crypto.DecrypterFactory;
-import org.dataportabilityproject.gateway.crypto.EncrypterFactory;
-import org.dataportabilityproject.gateway.crypto.SymmetricKeyGenerator;
+import org.dataportabilityproject.security.DecrypterFactory;
+import org.dataportabilityproject.security.EncrypterFactory;
+import org.dataportabilityproject.security.SymmetricKeyGenerator;
 import org.dataportabilityproject.gateway.reference.ReferenceApiUtils.FrontendConstantUrls;
 import org.dataportabilityproject.gateway.reference.ReferenceApiUtils.HttpMethods;
 import org.dataportabilityproject.spi.cloud.storage.JobStore;
